tasks/section-01: Reboot host after a dist-upgrade operation
Reboot the remote host after doing a dist-upgrade - since a few packages
such as kernel-headers, lisbssl, e.t.c take effect on reboot.

NOTE: I need a better method of checking whether a host need a reboot or
not, instead on relying on the changed status of apt dist-upgrade task.

Signed-off-by: James Oguya <oguyajames@gmail.com>