section-13: Split section 13.5 into two tasks
Instead of printing the command for auditing section 13.5 in a debug
msg, we can run it in one task, register the output in a variable and
in the subsequent task we can 'tell' the user to ensure that only the
word 'root' is returned!
Besides that, having special character ':' was messing up the ansible's
yaml parser and escaping it using quotes wasn't helping either.

This commit is related to issue #7
https://github.com/oguya/cis-ubuntu-14-ansible/issues/7

Signed-off-by: James Oguya <oguyajames@gmail.com>