section-5.1.7: Add default variable to allow user to override task
For common services such as dhcpd, tftp, x-windows, e.t.c you should
allow user to decide whether or not he/she wants to control the
effect/outcome of a task. To do this, we can use default variables,
which when initialized control the outcome of a task.

For instance: If disable_tftp_server is set to "yes", then tftp server
will be disabled by task with the tag section5.1.7.

Signed-off-by: James Oguya <oguyajames@gmail.com>