Section-07: Fix conflicting value in disable_ipv6 var.
In roles/default, we explicity set a value of 'no' to disable_ipv6
variable which means that we shouldn't disable IPv6. But in Section 07
task file, we are doing the exact opposite! Therefore, in the task file
we should check for a value of 'yes' in disable_ipv6 variable, and NOT
'no'.

Signed-off-by: James Oguya <oguyajames@gmail.com>